UBER-1157: Some dependant fixes
