+++ plugins/workbench-resources/src/components/Workbench.svelte
@@ -113,9 +113,9 @@
   let panelInstance: PanelInstance
   let popupInstance: Popup
 
-  let visibileNav: boolean = getMetadata(workbench.metadata.NavigationExpandedDefault) ?? true
+  let visibleNav: boolean = getMetadata(workbench.metadata.NavigationExpandedDefault) ?? true
   async function toggleNav (): Promise<void> {
-    visibileNav = !visibileNav
+    visibleNav = !visibleNav
     closeTooltip()
     if (currentApplication && navigatorModel && navigator) {
       await tick()
@@ -480,16 +480,16 @@
 
   let navFloat: boolean = !($deviceInfo.docWidth < 1024)
   $: if ($deviceInfo.docWidth <= 1024 && !navFloat) {
-    visibileNav = false
+    visibleNav = false
     navFloat = true
   } else if ($deviceInfo.docWidth > 1024 && navFloat) {
     if (getMetadata(workbench.metadata.NavigationExpandedDefault) === undefined) {
       navFloat = false
-      visibileNav = true
+      visibleNav = true
     }
   }
   const checkOnHide = (): void => {
-    if (visibileNav && $deviceInfo.docWidth <= 1024) visibileNav = false
+    if (visibleNav && $deviceInfo.docWidth <= 1024) visibleNav = false
   }
   let appsDirection: 'vertical' | 'horizontal'
   $: appsDirection = $deviceInfo.isMobile && $deviceInfo.isPortrait ? 'horizontal' : 'vertical'
@@ -598,7 +598,7 @@
     </clipPath>
   </svg>
   <div class="workbench-container" style:flex-direction={appsDirection === 'horizontal' ? 'column-reverse' : 'row'}>
-    <div class="antiPanel-application {appsDirection}" class:lastDivider={!visibileNav}>
+    <div class="antiPanel-application {appsDirection}" class:lastDivider={!visibleNav}>
       <div
         class="hamburger-container clear-mins"
         class:portrait={appsDirection === 'horizontal'}
@@ -617,8 +617,8 @@
         <div class="topmenu-container clear-mins flex-no-shrink" class:mini={appsMini}>
           <AppItem
             icon={TopMenu}
-            label={visibileNav ? workbench.string.HideMenu : workbench.string.ShowMenu}
-            selected={!visibileNav}
+            label={visibleNav ? workbench.string.HideMenu : workbench.string.ShowMenu}
+            selected={!visibleNav}
             size={appsMini ? 'small' : 'medium'}
             on:click={toggleNav}
           />
@@ -684,9 +684,9 @@
       }}
     />
     <div class="workbench-container">
-      {#if currentApplication && navigatorModel && navigator && visibileNav}
+      {#if currentApplication && navigatorModel && navigator && visibleNav}
         <!-- svelte-ignore a11y-click-events-have-key-events -->
-        {#if navFloat}<div class="cover shown" on:click={() => (visibileNav = false)} />{/if}
+        {#if navFloat}<div class="cover shown" on:click={() => (visibleNav = false)} />{/if}
         <div class="antiPanel-navigator {appsDirection === 'horizontal' ? 'portrait' : 'landscape'}">
           <div class="antiPanel-wrap__content">
             {#if currentApplication}
@@ -741,7 +741,7 @@
         }}
       >
         {#if currentApplication && currentApplication.component}
-          <Component is={currentApplication.component} props={{ currentSpace, visibileNav, navFloat, appsDirection }} />
+          <Component is={currentApplication.component} props={{ currentSpace, visibleNav, navFloat, appsDirection }} />
         {:else if specialComponent}
           <Component
             is={specialComponent.component}
@@ -749,7 +749,7 @@
               model: navigatorModel,
               ...specialComponent.componentProps,
               currentSpace,
-              visibileNav,
+              visibleNav,
               navFloat,
               appsDirection
             }}
@@ -764,7 +764,7 @@
         {:else if currentView?.component !== undefined}
           <Component
             is={currentView.component}
-            props={{ ...currentView.componentProps, currentView, visibileNav, navFloat, appsDirection }}
+            props={{ ...currentView.componentProps, currentView, visibleNav, navFloat, appsDirection }}
           />
         {:else}
           <SpaceView {currentSpace} {currentView} {createItemDialog} {createItemLabel} />
